+++ src/material-experimental/mdc-button/BUILD.bazel
@@ -2,7 +2,7 @@ package(default_visibility = ["//visibility:public"])
 
 load("@io_bazel_rules_sass//:defs.bzl", "sass_binary", "sass_library")
 load("//src/e2e-app:test_suite.bzl", "e2e_test_suite")
-load("//tools:defaults.bzl", "ng_e2e_test_library", "ng_module", "ng_test_library", "ng_web_test_suite", "ts_library")
+load("//tools:defaults.bzl", "ng_e2e_test_library", "ng_module", "ng_test_library", "ng_web_test_suite")
 
 ng_module(
     name = "mdc-button",
@@ -10,7 +10,6 @@ ng_module(
         ["**/*.ts"],
         exclude = [
             "**/*.spec.ts",
-            "harness/**",
         ],
     ),
     assets = [
@@ -23,18 +22,6 @@ ng_module(
         "//src/cdk/platform",
         "//src/material/core",
         "@npm//@material/ripple",
-    ],
-)
-
-ts_library(
-    name = "harness",
-    srcs = glob(
-        ["harness/**/*.ts"],
-        exclude = ["**/*.spec.ts"],
-    ),
-    deps = [
-        "//src/cdk/coercion",
-        "//src/cdk/testing",
     ],
 )
 
@@ -94,16 +81,17 @@ sass_binary(
 
 ng_test_library(
     name = "button_tests_lib",
-    srcs = [
-        "harness/button-harness.spec.ts",
-    ],
+    srcs = glob(
+        ["**/*.spec.ts"],
+        exclude = [
+            "**/*.e2e.spec.ts",
+        ],
+    ),
     deps = [
-        ":harness",
         ":mdc-button",
         "//src/cdk/platform",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
         "//src/material/button",
+        "//src/material/core",
         "@npm//@angular/platform-browser",
     ],
 )
