+++ src/material-experimental/mdc-button/testing/button-harness.ts
@@ -8,7 +8,7 @@
 
 import {ComponentHarness, HarnessPredicate} from '@angular/cdk/testing';
 import {coerceBooleanProperty} from '@angular/cdk/coercion';
-import {ButtonHarnessFilters} from './button-harness-filters';
+import {ButtonHarnessFilters} from '@angular/material/button/testing/button-harness-filters';
 
 
 /**
