+++ src/material/button/testing/button-harness.spec.ts
@@ -1,0 +1,7 @@
+import {MatButtonModule} from '@angular/material/button';
+import {runHarnessTests} from '@angular/material/button/testing/shared.spec';
+import {MatButtonHarness} from './button-harness';
+
+describe('Non-MDC-based MatButtonHarness', () => {
+  runHarnessTests(MatButtonModule, MatButtonHarness);
+});
